Added POC of a python client
This is a POC of how to make a client with the server API. The client
allow use always the newest config version but taking into account that
in one session the configurations obtained are consistent.

It's necessary refactor it to have a clean API, to have better documentation
and to be more solid but for a POC it's enough.