Initial version of a inotify updater
This updater use the inotify API to check the changes in the filesystem
and notify the mhconfig server about them. This is a POC and it don't
handle all the corner cases of the inotify API but it's a good starting
point. Some changes to do are:
- Handle better the file moves
- Review the mhconfig API for the updates (it has some limitations like
force reload all the files to handle the inotify start up or some crash
that makes it lose some changes)
- Check for corner cases