Added support for templates and removed support for partial get
A string template engine has been added to facilitate the server integration
with third party tools, and also with this change the support to obtain a
partial configuration has been removed because it don't make sense anymore.

Like this changes are to big I want to fix/polish some features like the
configuration optimization or the ability to specify the type of the
scalars and add in the template functions to convert the types.