Add support for types using the yaml tags
Now it's possible define the type of a scalar using the yaml tags
`!!str`, `!!int`, `!!float` and `!!bool`. This isn't neccesary if the
client know the type of the scalars but this could be a limitation
for the templates and for a generic client that convert the received
configuration to another format.