Use abseil for the map&set&mutex and fix some bugs
This changes mainly use the abseil library for the map&set&mutex, the reason
of this was in a first moment the performance and although it has been noted
that the mutex is somewhat slower the deadlock detection is an nice feature
(that has made it possible to detect a deadlock). Also some bugs and
minor changes as change a set to a vector has been done.