Added support for flavors and various enhancements
This is a big change that add support for flavors and for this the
internal way to store the documents has been changed flattening the
data structures and simplifying the overall logic. Also several fixes
has been done after testing and refactoring the changes, it would
look better in several commits but separating the changes would be
an arduous task and I don't think it's worth.