Added trace API call and refactor of the gRPC server
This changes add a new trace API call to investigate if some
document/template is still in use, who is using a certain configuration,
etc.

This has required a modification of the gRPC server wrapper since it
had an bug that caused memory leak if the connection was lost and
there was no event in the completion queue.