Changed overrides/flavors by labels
The overrides/flavors logic have the advantage that it is an easy algorithm to understand
but I am not convinced about its use since:

- I think the use of labels makes it easier to know what you're asking for.
- Practically the code of the overrides/flavors was duplicated.

Another thing is that I have separated a bit the code of the overrides generation from the
gRPC server one, so I could add a local mode to gather the configuration. This can be
interesting for development and for those people who prefer to have the configuration in
files, environment variables, etc. but want to have the power of an override system.

With this I want to use also the configuration system also for the daemon configuration,
for the moment this is only possible for the authentication configuration but in the future
I want to move all the daemon configuration there.