Refactor of the element class and improve the configuration debug
The element class internal structure has been changed to split the type
of the tag and also used the wasted bytes to store the position of the
value to improve the debug process.

With this change the error logic has been simplified with a logger
interface that allow send to the clients the build process to detect
error. A simple python checker has been added has a POC to use this
debug logic (it could be found in the path `client/python`).

Consequently, error handling has been improved, sending at least a
partial config.